Move calls to restraints web service into app/scenarios/restraints.ts
